♻️ Move the setup flow out of features
Setup is a route, not a feature: its form, footer, and space action now
live in app/[locale]/setup, and isInitialAdmin moves to the
instance-settings feature it actually belongs to. Also refines the
footer copy (medium weight email, "Sign out?") under fresh i18n keys,
autofocuses the organization name field when Work is selected, and uses
the small logo in the header.
