+++ apps/web/public/locales/en/app.json
@@ -627,7 +627,8 @@
   "settings": "Settings",
   "setupAccountDescription": "Tell us a bit about yourself.",
   "setupAccountTitle": "Set Up Your Account",
-  "setupLoggedInAs": "Logged in as {email}",
+  "setupFooterLoggedInAs": "Logged in as <email>{email}</email>",
+  "setupFooterSignOut": "Sign out?",
   "setupPasswordAlertDesc": "Set up a password to make logging in faster and more convenient.",
   "setupPasswordAlertTitle": "Login quicker with a password",
   "setupPasswordButton": "Set up password",
