+++ apps/web/src/app/[locale]/setup/actions.ts
