+++ apps/web/src/app/[locale]/setup/components/setup-footer.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { Button } from "@rallly/ui/button";
 import React from "react";
 import { Trans } from "@/i18n/client";
 import { signOut } from "@/lib/auth-client";
@@ -8,16 +9,21 @@ export function SetupFooter({ email }: { email: string }) {
   const [isSigningOut, setIsSigningOut] = React.useState(false);
 
   return (
-    <p className="text-muted-foreground text-sm">
-      <Trans
-        i18nKey="setupLoggedInAs"
-        defaults="Logged in as {email}"
-        values={{ email }}
-      />{" "}
-      <button
-        type="button"
-        disabled={isSigningOut}
-        className="text-link disabled:opacity-50"
+    <p className="flex flex-col items-center gap-4 text-muted-foreground text-sm">
+      <span>
+        <Trans
+          i18nKey="setupFooterLoggedInAs"
+          defaults="Logged in as <email>{email}</email>"
+          values={{ email }}
+          components={{
+            email: <span className="font-medium text-foreground" />,
+          }}
+        />
+      </span>
+      <Button
+        variant="ghost"
+        size="sm"
+        loading={isSigningOut}
         onClick={() => {
           setIsSigningOut(true);
           signOut()
@@ -30,8 +36,8 @@ export function SetupFooter({ email }: { email: string }) {
             });
         }}
       >
-        <Trans i18nKey="logout" defaults="Logout" />
-      </button>
+        <Trans i18nKey="signOut" defaults="Sign out" />
+      </Button>
     </p>
   );
 }
